.automation/requirements.txt
rich==13.7.1
typer==0.12.3
pyyaml==6.0.1
requests==2.31.0
jsonschema==4.22.0
pillow==10.3.0
numpy==1.26.4
